refactored to make things more modular (I was failing windows checks)
