+++ harden/add_local_admin_linux.go
@@ -9,17 +9,6 @@ import (
 	"github.com/spf13/cobra"
 )
 
-var addLocalAdminCmd = &cobra.Command{
-	Use:   "add-admin [username]",
-	Short: "Adds a local admin to this system",
-	Long:  `Adds a local admin and prompts the user to enter the password.`,
-	Args:  cobra.ExactArgs(1),
-	Run: func(cmd *cobra.Command, args []string) {
-		username := args[0]
-		addLocalAdmin(username)
-	},
-}
-
 func addLocalAdmin(username string) {
 	cmd := exec.Command("useradd", username, "--no-create-home", "--home-dir", "/")
 	if err := runCommand(cmd); err != nil {
@@ -55,8 +44,4 @@ func runCommand(cmd *exec.Cmd) error {
 		fmt.Printf("Stderr: %s\n", stderr.String())
 	}
 	return err
-}
-
-func setupAddLocalAdminCmd(cmd *cobra.Command) {
-	cmd.AddCommand(addLocalAdminCmd)
 }
